+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## v8.7.0
-* Updated the Cast condition and icon type to use LibClassicCasterino for approximations of others' spell casts.
+* Classic: Updated the Cast condition and icon type to use LibClassicCasterino for approximations of others' spell casts.
 
 ### Bug Fixes
 * Classic: Fixed the Unit Class condition's options.
+* Classic: Fixed the Weapon Imbue icon type & Condition for offhands.
 
 ## v8.6.9
 * Classic: Aura durations now account for learned talents.