+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v8.7.0
+* The Missing Buffs/Debuffs icon type now sorts by lowest duration first.
+
 ## v8.6.9
 ### Bug Fixes
 * Fixed an issue with Unit Conditions where the initial state of the conditions sometimes wouldn't be taken into account.